handle comment marker from plugin
