+++ src/lib/comments.ts
@@ -22,7 +22,6 @@ import { NodeWithPos } from 'prosemirror-utils'
 import { EditorView } from 'prosemirror-view'
 
 import { addNodeComment } from '../commands'
-import { commentsKey, setCommentSelection } from '../plugins/comments'
 
 export type CommentAttrs = CommentNode['attrs']
 export type HighlightMarkerAttrs = HighlightMarkerNode['attrs']
@@ -88,12 +87,6 @@ export const getCommentRange = (comment: CommentAttrs) => {
   }
 }
 
-export const handleComment = (node: ManuscriptNode, view: EditorView): void => {
-  const { state } = view
-
-  addNodeComment(node, state, view.dispatch)
-}
-
 export const createCommentMarker = (
   tagName: string,
   key: CommentKey,
@@ -123,40 +116,8 @@ const getMarkerID = (id: string) => {
   return `${id}-comment-marker`
 }
 
-export const handleCommentMarkerClick = (event: Event, view: EditorView) => {
-  const element = event.target as HTMLElement
-  // Handle click on comment marker
-  const marker = element.closest('.comment-marker') as HTMLElement
-  if (marker) {
-    const key = marker.dataset.key as CommentKey
-    const tr = view.state.tr
-    setCommentSelection(tr, key, undefined, false)
-    view.dispatch(tr)
-    return
-  }
-}
-export const updateCommentSelection = (
-  marker: HTMLElement,
-  view: EditorView
-) => {
-  const key = marker.dataset.key as CommentKey
-  const com = commentsKey.getState(view.state)
+export const handleComment = (node: ManuscriptNode, view: EditorView): void => {
+  const { state } = view
 
-  const comments = com?.commentsByKey.get(key)
-  if (!comments) {
-    marker.setAttribute('data-count', '0')
-  } else if (comments.length !== 1) {
-    marker.setAttribute('data-count', String(comments.length))
-  } else {
-    marker.removeAttribute('data-count')
-  }
-
-  const isSelected = key === com?.selection?.key
-  if (isSelected) {
-    marker.classList.add('selected-comment')
-  } else {
-    marker.classList.remove('selected-comment')
-  }
-
-  return isSelected
+  addNodeComment(node, state, view.dispatch)
 }
