+++ styles/AdvancedEditor.css
@@ -249,7 +249,12 @@
   cursor: pointer;
   position: relative;
 }
-
+#editor .manuscript-title ~.comment-marker {
+  height: 0;
+  top: -13px;
+  left: 60px;
+  display: block;
+}
 .block:not(.affiliations) > div.comment-marker,
 .figure-block > div.comment-marker {
   top: -25px;
